pyproject.toml
[project]
name = "spotipython"
version = "0.0.1"
description = "Autogenerated Spotify Web API client library. Supports async and typed."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.3.0",
    "httpx>=0.28.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
]

authors = [
    { name="Maurice Banerjee Palmer", email="spotipython@maurice.fm"},
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

[dependency-groups]
dev = [
    "openapi-python-client>=0.26.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/mbanerjeepalmer/spotipython"



[build-system]
# TODO is this the correct version constraint?
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "spotipython"
module-root = ""
